Normalize image rotation in iOS for ML vision input
It's common in iOS to see an image rotated, e.g. a photo shot in
portrait mode will have 90 degree rotated image, its orientation
is going to be `UIImageOrientationRight`.

It makes difficult to map text detection result rectangles for
example.
So it comes the image rotation normalization. Yet the process is not
performance wise but it works at least.